move logging to separate file
